# Conflicts:
#	test/legacy_test/test_api_compatibility_part5.py

+++ paddle/phi/backends/gpu/gpu_launch_config.h
@@ -143,14 +143,14 @@ inline GpuLaunchConfig GetGpuLaunchConfig1D(const GPUContext& dev_ctx,
   // Number of threads per block shall be larger than 64.
   threads = std::max(64, threads);
   int64_t blocks = DivUp<int64_t>(DivUp<int64_t>(numel, vec_size), threads);
-  int limit_blocks = dev_ctx.GetCUDAMaxGridDimSize()[0];
+  int64_t limit_blocks = dev_ctx.GetCUDAMaxGridDimSize()[0];
   if (blocks > limit_blocks) {
     blocks = limit_blocks;
   }
 
   GpuLaunchConfig config;
   config.thread_per_block.x = threads;
-  config.block_per_grid.x = blocks;
+  config.block_per_grid.x = static_cast<uint32_t>(blocks);
   config.compute_capability = capability;
 
   VLOG(7) << "Get 1-D launch config: numel=" << numel
