+++ paddle/phi/backends/onednn/onednn_helper.h
@@ -130,20 +130,20 @@ inline dnnl::memory::data_type OneDNNGetDataType<dtype::bfloat16>() {
 inline std::vector<std::vector<int64_t>> ToOneDNNPadding(
     const std::vector<int64_t>& paddings) {
   if (paddings.size() == 6) {
-    int padding_front = paddings[0];
-    int padding_back = paddings[1];
-    int padding_top = paddings[2];
-    int padding_bottom = paddings[3];
-    int padding_left = paddings[4];
-    int padding_right = paddings[5];
+    int64_t padding_front = paddings[0];
+    int64_t padding_back = paddings[1];
+    int64_t padding_top = paddings[2];
+    int64_t padding_bottom = paddings[3];
+    int64_t padding_left = paddings[4];
+    int64_t padding_right = paddings[5];
 
     return {{padding_front, padding_top, padding_left},
             {padding_back, padding_bottom, padding_right}};
   } else {
-    int padding_top = paddings[0];
-    int padding_bottom = paddings[1];
-    int padding_left = paddings[2];
-    int padding_right = paddings[3];
+    int64_t padding_top = paddings[0];
+    int64_t padding_bottom = paddings[1];
+    int64_t padding_left = paddings[2];
+    int64_t padding_right = paddings[3];
 
     return {{padding_top, padding_left}, {padding_bottom, padding_right}};
   }
